+++ src/views/admin/EditCreateAccountForm.vue
@@ -40,7 +40,6 @@
             <student-information-section
                 v-if="isStudent"
                 v-model:matriculation-id="account.student.matriculationId"
-                v-model:immatriculation-has-change="immatriculationHasChange"
                 :edit-mode="editMode"
                 :error-bag="errorBag"
                 :username="account.user.username"
@@ -199,7 +198,6 @@
             let success = ref(false);
             let deleteModal = ref();
             let unsavedChangesModal = ref();
-            let immatriculationHasChange = ref(false);
 
             const errorBag = ref(new ErrorBag());
 
@@ -297,8 +295,7 @@
                     account.lecturer.freeText != initialAccount.lecturer.freeText ||
                     account.lecturer.researchArea != initialAccount.lecturer.researchArea ||
                     //student properties
-                    account.student.matriculationId != initialAccount.student.matriculationId ||
-                    immatriculationHasChange.value
+                    account.student.matriculationId != initialAccount.student.matriculationId
                 ) {
                     emit("update:has-input", true);
                     return true;
@@ -429,7 +426,6 @@
                 deleteModal,
                 unsavedChangesModal,
                 errorBag: errorBag,
-                immatriculationHasChange,
             };
         },
     };