+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Bugfix
 - Error feedback is given on posting an account without role [#535](https://github.com/upb-uc4/ui-web/issues/535)
 
+# [v.0.9.0-hotfix.1](https://github.com/upb-uc4/ui-web/compare/v0.9.0...v0.9.0-hotfix.1) (2020-09-28)
+
+## Bugfix
+- enable error display for adding matriculation data [#543](https://github.com/upb-uc4/ui-web/pull/543)
+
 # [v.0.9.0](https://github.com/upb-uc4/ui-web/compare/v0.8.1...v0.9.0) (2020-09-28)
 
 ## Feature